Refactor upgrade helpers into shared module
